리스트,디테일
